Fix std::bad_alloc error thrown in GstVideoPlayer constructor
width_ and height_ were not being assigned a default value - and the call to GetVideoSize often fails to get a valid structure.
This results in the width_ and height_ being uninitialized by the time they are used to set the internal video size. Hence it ends up allocating a random amount of memory which can be greater than the available system memory - crashing the plugin and flutter wrapper.